requirements.txt
Flask==1.1.2
gunicorn==19.7.1
pyjwt==1.7.1
